fix: allow NetworkIn only with public IPv4 addresses
