+++ util/netutil/net_test.go
@@ -50,3 +50,21 @@ func TestIsPrivateIPv6(t *testing.T) {
 		assert.Equal(t, b, cc.isPriv, cc.ip)
 	}
 }
+
+func TestSortedIPs(t *testing.T) {
+	ips := []string{
+		"192.168.70.17",
+		"46.148.198.133",
+		"fd21:f7bb:61b8:9e37::1",
+		"2001:db8::68",
+	}
+
+	sortedIPs := []string{
+		"2001:db8::68",
+		"46.148.198.133",
+		"fd21:f7bb:61b8:9e37::1",
+		"192.168.70.17",
+	}
+
+	assert.Equal(t, sortedIPs, SortedIPs(ips))
+}
